typos + added a bunch of comments
Please take a look at figures - some subpanels need to be better labeled and some are missing additional text descriptions. 

One of the last paragraphs of Results is completely scrambled and I cannot figure out what you tried to say.

Added some space btw paragraphs so that it's easier to read. 

Make sure that the figures and panels are introduced in the text as they appear.